% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlsData.R
\name{getBlsData}
\alias{getBlsData}
\title{getBlsData}
\usage{
getBlsData(start, seriesID)
}
\arguments{
\item{start}{starting year of the time series data}

\item{seriesID}{BLS series ID number for the time series data}
}
\value{
data frame with 10 years of data
}
\description{
getBlsData
}
